A Shell program.
This was a semester-long group project for a Systems Programming course.
I, along with 3 other group members, designed and implemented a Shell
with several functions such as: the ability to do background jobs with
the & symbol, a Word Count program, the ability to pipe, and all the
basic commands all Shells have.